+++ controllers/termsofuse.js
@@ -1,10 +1,9 @@
 const express = require('express');
 const { URL } = require('url');
-const api = require('../api');
-const authHelper = require('../helpers/authentication');
 const { DOCUMENT_BASE_DIR, SC_THEME } = require('../config/global');
 const { specificFiles } = require('../config/documents');
 const { getBase64File } = require('../helpers/fileHelper');
+const { getConsentVersion } = require('../helpers/consentVersionHelper');
 
 const router = express.Router();
 
@@ -12,20 +11,7 @@ const termsUrl = () => new URL(`${SC_THEME}/${specificFiles.termsOfUseSchool}`, 
 
 router.get('/', async (req, res, next) => {
 	try {
-		const isAuthenticated = await authHelper.isAuthenticated(req);
-		const qs = {
-			$limit: 1,
-			consentTypes: ['termsOfUse'],
-			$sort: {
-				publishedAt: -1,
-			},
-		};
-
-		if (isAuthenticated && res.locals.currentSchool) {
-			qs.schoolId = res.locals.currentSchool;
-		}
-
-		const consentVersions = await api(req).get('/consentVersions', { qs });
+		const consentVersions = await getConsentVersion(req, res, 'termsOfUse');
 
 		if (consentVersions.data.length) {
 			const fileId = consentVersions.data[0].consentDataId;
